==> notebooks/2_Data_Exploration.ipynb <==
# 2. Data cleaning and initial exploration

In this notebook we will be loading and performing an initial exploration of the dataset from the Protocols track for the Hercules challenge. This dataset consists of a series of 100 experimental protocols from the [Bio-Protocols journal](https://bio-protocol.org/Default.aspx).

## Setup
As always, we will begin by loading the logging system and setting up the path to import additional code from the _src_ folder.

%run __init__.py

from bokeh.io import output_notebook

output_notebook()

from herc_common import BokehHistogram

hist = BokehHistogram(color_fill="mediumslateblue", color_hover="slateblue")

def print_empty_cols(df):
    for col in df.columns:
        print(col)
        print('-' * len(col))
        res = df[df[col] == ''].index
        print(f"{len(res)} articles have no value for column {col}")
        print(res)
        print('\n')


## Parsing the data
Since all the _html_ files of each protocol have been scraped in the previous phase, we will begin by getting the path of all the files that will be parsed later on:

import glob

files_to_read = glob.glob(f"{PROTOCOLS_DIR}/*.html")
len(files_to_read)

We will also define a simple Protocol class that will be used to stored the parsed information from every HTML file, and provides some utility methods to be loaded in a pandas DataFrame:

from src.protocol import Protocol

Since not every protocol will have a value for every defined field (for example, some protocols may not have a background section), we will also define a decorator that will be used later on to specify which fields are optional. These fields will be an empty string if no value can be found for them:

from src.data_reader import parse_protocol

protocols = []
for file in files_to_read:
    protocol_id = os.path.basename(file).split('.')[0]
    with open(file, 'r', encoding='utf-8') as f:
        protocols.append(parse_protocol(f.read(), protocol_id))

## Creating a dataframe

Now that every protocol has been parsed, we can load them into a pandas DataFrame:

import pandas as pd

df = pd.DataFrame([protocol.to_dict() for protocol in protocols])
df.head()

We can already see that none of the first 4 protocols have a value for the "background" field. we will analyse the data more in depth in the following sections.

## Cleaning and feature engineering

df.describe()

From the output above we can see that there is a unique id, title, abstract, procedure and equipment for each different protocol. Other fields have some values which are not unique. This could mean either that they have the same value or that they are empty. We will check later on whether they are empty or not.

For now, we will start by joining all the steps of each procedure into a new column which will be called '_full\_procedure\_cleaned_':

import re

def join_procedure_steps(procedure):
    return ' '.join(procedure.split('|'))

def clean(procedure):
    merged_procedure = join_procedure_steps(procedure)
    return re.sub('\s+', ' ', merged_procedure).strip()
    

df['full_text'] = df['title'] + '. ' + df['abstract'] + '. ' + df['procedure'] + '. ' + df['background']
df['full_text_no_abstract'] = df['title'] + '. ' + df['procedure'] + '. ' + df['background']
df['full_text_cleaned'] = df['full_text'].apply(lambda x: clean(x))
df['full_text_no_abstract_cleaned'] =df['full_text_no_abstract'].apply(lambda x: clean(x))
df['full_text_cleaned'].loc[0][:500]

We will finally add another column with the number of characters of each procedure:

df['num_chars_text'] = df['full_text_cleaned'].apply(lambda x: len(x))

## Initial exploration

As we have seen before, some of the values for the background, materials, categories and authors columns are not unique. We will se now whether they contain empty values or not:

df.head()

print_empty_cols(df)

As we can see, for the materials and background columns this is the case (all non-unique values are empty ones). However, there are repeated values for the equipment and authors fields, which is expected.

If we are going to use the bakground column in future steps we have to notice that most of the protocols don't have a value for it.

We are going to see how the length of each procedure is distributed amongst the dataset:

df['num_chars_text'].describe()

HIST_COLUMN = 'num_chars_text'
HIST_TITLE = "Procedure length distribution for the Bio-Protocols dataset"
HIST_XLABEL = "Procedure length (# of characters)"
HIST_YLABEL = "Number of protocols"

hist.load_plot(df, HIST_COLUMN, HIST_TITLE,
               HIST_XLABEL, HIST_YLABEL, True)

df.loc[0].full_text_cleaned

From the cells above we can see that the mean length of each procedure is about 5800 characters, and most of the protocols have procedures of 2000-7000 characters.

hist.save_plot(os.path.join(NOTEBOOK_2_RESULTS_DIR, '1_Protocol_procedure_length.svg'))

## Saving the dataframe

Finally, we are going to save the dataframe so we can use it later on in the following steps:

DF_FILE_PATH = os.path.join(NOTEBOOK_2_RESULTS_DIR, 'protocols_dataframe.pkl')

df.to_pickle(DF_FILE_PATH)
