==> notebooks/6_Complete_System.ipynb <==
# 6. Complete System
In this notebook we will be joining the pieces developed in the previous notebooks to create the final pipeline used to obtain the topics of a given publication.

In this notebook we will be working on the Topic Combination module that will combine the lists of potential topics to output the final list of topics returned by the system with their confidence scores. Once this module is finished, the complete system will be used to obtain the topics of each article in the dataset, and we will save it for later use and inferring topics from new data.

# Setup

%run __init__.py

import pandas as pd

DF_FILE_PATH = os.path.join(NOTEBOOK_2_RESULTS_DIR, 'protocols_dataframe.pkl')

df = pd.read_pickle(DF_FILE_PATH)
protocols = df['full_text_cleaned'].values

df.head()

## Loading the model
The main two pipelines that retrieve the list of topics from the text will be loaded here:

from herc_common.utils import load_object

lda_pipe = load_object(os.path.join(NOTEBOOK_4_RESULTS_DIR, 'lda_pipe_with_labels.pkl'))
ner_pipe = load_object(os.path.join(NOTEBOOK_5_RESULTS_DIR, 'topic_extraction_from_ner_pipe.pkl'))

## Combining topics
To join the results of both pipelines we will be making use of the [FeatureUnion](https://scikit-learn.org/stable/modules/generated/sklearn.pipeline.FeatureUnion.html) class from scikit-learn. This class will serve as a single transformer which concatenates the results of the previous pipelines:

from sklearn.pipeline import FeatureUnion

union = FeatureUnion([("ner", ner_pipe),
                      ("lda", lda_pipe)])

## Building the final pipeline
To build our final system, we will make use of a custom class that will combine the topics from the _ner_ and _lda_ pipelines. Once the topics are concatenated by the feature union they will go to the combiner, where a final list of topics will be returned:

from sklearn.pipeline import Pipeline

from herc_common.topic import TopicCombiner


combiner = TopicCombiner()
final_pipe = Pipeline([('union', union),
                      ('combiner', combiner)])

import string

import en_core_sci_lg
import en_core_web_md

from collections import Counter

from tqdm import tqdm

en_core_web_md.load()
en_core_sci_lg.load()

final_pipe.transform([protocols[-1]])

## Predicting the final topics for the dataset
Now that we have our final system ready, we will obtain the list of topics for the Agriculture dataset:

topics = final_pipe.transform(protocols)

topics[0][:5]

## Saving results
Finally, we are going to save the complete pipeline for further use with new data:

from herc_common.utils import save_object

save_object(final_pipe, os.path.join(NOTEBOOK_6_RESULTS_DIR, 'final_pipe.pkl'))

results_df = df[['pr_id', 'title', 'categories']]
results_df = results_df.assign(y_pred=topics)
results_df.head()

results_df.to_csv(os.path.join(NOTEBOOK_6_RESULTS_DIR, 'protocol_results.csv'), index=False)
