==> notebooks/7_Evaluation.ipynb <==
# 7. Evaluation of results
In this final notebook we will propose an evaluation metric to check the performance and our final systems on the protocols track. For our approach we will rely on the default categories asigned to each protocols. For an additional level of accuracy in the evaluation of our models, the list of subjects used as ground truth could be generated by a human committee. For the scope of this challenge, we will however rely on the default subjects present in each protocol and leave the human annotation as future work.

## Setup

%run __init__.py

import pandas as pd

PROTOCOLS_FILE_PATH = os.path.join(NOTEBOOK_2_RESULTS_DIR, 'protocols_dataframe.pkl')

protocols_df = pd.read_pickle(PROTOCOLS_FILE_PATH)

## Analyzing the categories
We will begin by selecting a subset of the protocols dataframe with just the id of the protocol and its categories:

import numpy as np

categories_df = protocols_df.copy()
categories_df['categories'].replace('', np.nan, inplace=True)
categories_df.dropna(subset=['categories'], inplace=True)
categories_df.head(n=4)

protocols_categories = {aid: categories.split('|') 
                       for aid, categories in zip(categories_df['pr_id'].values,
                                                  categories_df['categories'].values)}
protocols_categories['e1029']

As we can see above, each protocol is composed of a variable sized list of category terms.

In the following cell we are going to perform a cleaning of the categories to remove those that will not be useful for the evaluation of our models. We will remove those words that do not have a match to WordNet, which will be used later on to perform the evaluation of the models. Finally, those rows that do not have any caetgory will be removed from the final sample:

from herc_common.evaluation import _get_synset

stop_categories = set([])

filtered_protocols_categories = {
    k: set([el for el in v if el.lower() not in stop_categories
            and not el.isnumeric() and _get_synset(el) is not None])
    for k, v in protocols_categories.items()
}

final_protocols_categories = {
    k: v
    for k, v in filtered_protocols_categories.items()
    if len(v) != 0
}

len(final_protocols_categories)

final_protocols_categories

As we can see above, from the initial 100 protocols in the dataframe 96 have at least a category to be compared against.

## Evaluation
For the evaluation of our system we will use WordNet to obtain a semantic similarity score between the topics predicted by our system and those used as ground truth.

Before we can start calculating these similarity scores, we will obtain the topics predicted by our system. First, we will be loading the final pipeline that has been saved in our previous notebook:

import string

import en_core_sci_lg
import en_core_web_md

from collections import Counter

from tqdm import tqdm

en_core_web_md.load()
en_core_sci_lg.load()

from herc_common.utils import load_object

final_pipe = load_object(os.path.join(NOTEBOOK_6_RESULTS_DIR, 'final_pipe.pkl'))

Now, we will select the sample of publications with at least one ground truth subject, and obtain the output of our system for those articles:

protocols_keys = final_protocols_categories.keys()
X = categories_df.set_index('pr_id', inplace=False).loc[protocols_keys]['full_text_cleaned'].values

y_base = final_protocols_categories.values()
y_pred = final_pipe.transform(X)

y_pred = [[str(topic[0]) for topic in doc] for doc in y_pred]
y_pred[:5]

## Similarity
In this section we will be calculating the similarity scores between the topics inferred by the model and the ones used as ground truth:

from herc_common.evaluation import compute_similarity_scores

scores = [compute_similarity_scores(y_b, y_p, 'lch_similarity')
          for y_b, y_p in zip(y_base, y_pred)]
scores[:5]

final_similarity = np.mean([score['mean similarity'] for score in scores])
final_similarity

## Saving the results
Finally, we are going to save the results. First of all, the predictions will be saved to a new dataframe:

cols_subset = ['title', 'categories']

results_df = categories_df.set_index('pr_id', inplace=False).loc[protocols_keys][cols_subset]
results_df['Topics Predicted'] = ['\n'.join(topics) for topics in y_pred]
results_df['categories'] = ['\n'.join(topics) for topics in y_base]
results_df.head()

scores_df = pd.DataFrame.from_records(scores)
scores_df.set_index(results_df.index, inplace=True)
scores_df.head()

And now the scores obtained for each protocol will be saved too:

final_df = results_df.join(scores_df)
final_df.head()

final_df.to_csv(os.path.join(NOTEBOOK_7_RESULTS_DIR, 'protocols_scores.csv'))
